+++ packages/plugin-core/src/test/suite-integ/WindowWatcher.test.ts
@@ -116,7 +116,7 @@ suite("WindowWatcher: GIVEN the dendron extension is running", function () {
         },
       },
       () => {
-        test("THEN preview panel is shown", async () => {
+        test.only("THEN preview panel is shown", async () => {
           const mockExtension = new MockDendronExtension({
             engine: ExtensionProvider.getEngine(),
             wsRoot: ExtensionProvider.getDWorkspace().wsRoot,
